+++ src/impls.rs
@@ -4,6 +4,7 @@ mod create;
 mod extend;
 mod index;
 mod intersection;
+mod into_iter;
 mod sub_iterable;
 mod sub_self;
 mod union;
